Add support for metaport config splitting
