+++ .gitignore
@@ -24,7 +24,7 @@ yarn-debug.log*
 yarn-error.log*
 
 .env
-src/data/metaportConfig*.ts
+src/data/metaportConfig
 src/meta/
 src/assets/validators/index.ts
 src/metadata.json
