+++ config/mainnet/calypso.ts
@@ -1,0 +1,110 @@
+import { type types } from '@/core'
+
+const connections: types.mp.TokenTypeMap = {
+    eth: {
+        eth: {
+            address: '0x59ab97Ee239e02112652587F9Ef86CB6F762983b',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                },
+                'green-giddy-denebola': {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        }
+    },
+    erc20: {
+        skl: {
+            address: '0x4048C4dd6eccF1Dc23b068211fDf20AD19602e50',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        usdt: {
+            address: '0xf9B5736E74b37Ec2EDd37B1B3c8e1aF9fa426F2A',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                },
+                'green-giddy-denebola': {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        dai: {
+            address: '0x941C0Dd6055787aE9D417620E8F4F540d8B9D9E5',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        usdc: {
+            address: '0x7Cf76E740Cb23b99337b21F392F22c47Ad910c67',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                },
+                'green-giddy-denebola': {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                },
+                'parallel-stormy-spica': {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        wbtc: {
+            address: '0x98f7d76CfFaD0263Ddb30B10244C02530Ba704c4',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        skivvy: {
+            address: '0x0d3FaC688AD94f01B237BFD795C61eDE6Ac9542E',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+    }
+}
+
+export default connections
