+++ config/mainnet/europa.ts
@@ -1,0 +1,238 @@
+import { type types } from '@/core'
+
+const connections: types.mp.TokenTypeMap = {
+    eth: {
+        eth: {
+            address: '0xD2Aaa00700000000000000000000000000000000',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'honorable-steel-rasalhague': {
+                    wrapper: '0xa5274efA35EbeFF47C1510529D9a8812F95F5735'
+                },
+                'green-giddy-denebola': {
+                    wrapper: '0xa5274efA35EbeFF47C1510529D9a8812F95F5735'
+                }
+            }
+        }
+    },
+    erc20: {
+        flag: {
+            address: '0xcdF030a3E65f917DFa8d74555A64a5eC5303c88e',
+            chains: {
+                'green-giddy-denebola': {}
+            }
+        },
+        reth: {
+            address: '0x4c21E0Bde3f02aDea7a874DBd705c30a3187eeaf',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'green-giddy-denebola': {
+                    wrapper: '0x113b3B0B813B856fbfC366595EdD85bC7A4DB74A'
+                }
+            }
+        },
+        skl: {
+            address: '0xE0595a049d02b7674572b0d59cd4880Db60EDC50',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'honorable-steel-rasalhague': {
+                    wrapper: '0xD162bB5c75FE99144295b03510bAb2DF99617440'
+                },
+                'green-giddy-denebola': {
+                    wrapper: '0xD162bB5c75FE99144295b03510bAb2DF99617440'
+                },
+                'affectionate-immediate-pollux': {
+                    wrapper: '0xD162bB5c75FE99144295b03510bAb2DF99617440'
+                }
+            }
+        },
+        wbtc: {
+            address: '0xcb011E86DF014a46F4e3AC3F3cbB114A4EB80870',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'green-giddy-denebola': {
+                    wrapper: '0x497C476bcABB2F7D50afF99ca3Ae9354451Ff7d0'
+                },
+                'honorable-steel-rasalhague': {
+                    wrapper: '0x497C476bcABB2F7D50afF99ca3Ae9354451Ff7d0'
+                }
+            }
+        },
+        usdt: {
+            address: '0x1c0491E3396AD6a35f061c62387a95d7218FC515',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'honorable-steel-rasalhague': {
+                    wrapper: '0x42dDeBDa2195FaEDAEC1cf24c19B42da1c881feB'
+                },
+                'green-giddy-denebola': {
+                    wrapper: '0x42dDeBDa2195FaEDAEC1cf24c19B42da1c881feB'
+                }
+            }
+        },
+        usdp: {
+            address: '0x73d22d8a2D1f59Bf5Bcf62cA382481a2073FAF58',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        ruby: {
+            address: '0x2B4e4899b53E8b7958c4591a6d02f9C0b5c50F8f',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        sushi: {
+            address: '0xcec20381ac92d610ac444f6d73c1d40726650259',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        link: {
+            address: '0x688B6B03E3ff9E9b38a9a05cCDC653D37e13fbE0',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        pepe: {
+            address: '0x95DEFbe12A9ef1A2B5372cbDd8aDd5ddC29CF247',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        cgt: {
+            address: '0xf97a23a0E49e5E0028F8Ce89F7be9a9FD5942Cf6',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        wct1: {
+            address: '0xF9c46D87b25E772BEaA2349EA49C46d7CB4d2B29',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        dai: {
+            address: '0xD05C4be5f3be302d376518c9492EC0147Fa5A718',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'honorable-steel-rasalhague': {
+                    wrapper: '0x9E7563B0D51F9B6dFB1917F97D29D50804300d28'
+                }
+            }
+        },
+        usdc: {
+            address: '0x5F795bb52dAC3085f578f4877D450e2929D2F13d',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'honorable-steel-rasalhague': {
+                    wrapper: '0x1c566a47e1baC535Ca616373146e3BE024F88Aa4'
+                },
+                'green-giddy-denebola': {
+                    wrapper: '0x1c566a47e1baC535Ca616373146e3BE024F88Aa4'
+                },
+                'parallel-stormy-spica': {
+                    wrapper: '0x1c566a47e1baC535Ca616373146e3BE024F88Aa4'
+                }
+            }
+        },
+        hmt: {
+            address: '0xBE3530a3eDf9472693065041B8c9155C7FeCB8e5',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'wan-red-ain': {
+                    wrapper: '0xA0f4D4db1457E442b83555cb92aaBB8DE959Aa75'
+                }
+            }
+        },
+        skill: {
+            address: '0xBDDad45160E10C3738785d9dD7F30b4B2a5Eeba8',
+            chains: {
+                'affectionate-immediate-pollux': {
+                    clone: true
+                }
+            }
+        },
+        cmps: {
+            address: '0x31E0B85Eff93fc3cb5423cd3E23868F5B18fF0bB',
+            chains: {
+                'green-giddy-denebola': {}
+            }
+        },
+        exd: {
+            address: '0xCfEBA92BD362B2F76fC30a89C433DE50a1D62BcA',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        hmkr: {
+            address: '0x57a3d5faa172f5984cdcd4885fd4f0baac88e558',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        paxg: {
+            address: '0x18478516f526b7947d995e04871562adbeeaa62d',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        razor: {
+            address: '0xCA46B70cA3c510Ce9D0c43D25817032e2F5354c0',
+            chains: {
+                mainnet: {
+                    clone: true
+                }
+            }
+        },
+        skivvy: {
+            address: '0xE0a320b0d525BA7a97afcE932F78789Db23c0e4a',
+            chains: {
+                mainnet: {
+                    clone: true
+                },
+                'honorable-steel-rasalhague': {
+                    wrapper: '0xD6b78761557E50BC94B7a99d75Fa6C9c29ab77e1'
+                }
+            }
+        },
+    }
+}
+
+export default connections
