+++ config/mainnet/mainnet.ts
@@ -1,0 +1,179 @@
+import { type types } from '@/core'
+
+const connections: types.mp.TokenTypeMap = {
+    eth: {
+        eth: {
+            chains: {
+                'elated-tan-skat': {},
+                'adorable-quaint-bellatrix': {},
+                'honorable-steel-rasalhague': {
+                    hub: 'elated-tan-skat'
+                },
+                'green-giddy-denebola': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        }
+    },
+    erc20: {
+        reth: {
+            address: '0xae78736cd615f374d3085123a210448e74fc6393',
+            chains: {
+                'elated-tan-skat': {},
+                'green-giddy-denebola': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        skl: {
+            address: '0x00c83aeCC790e8a4453e5dD3B0B4b3680501a7A7',
+            chains: {
+                'elated-tan-skat': {},
+                'honorable-steel-rasalhague': {
+                    hub: 'elated-tan-skat'
+                },
+                'green-giddy-denebola': {
+                    hub: 'elated-tan-skat'
+                },
+                'affectionate-immediate-pollux': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        wbtc: {
+            address: '0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599',
+            chains: {
+                'elated-tan-skat': {},
+                'green-giddy-denebola': {
+                    hub: 'elated-tan-skat'
+                },
+                'honorable-steel-rasalhague': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        usdt: {
+            address: '0xdAC17F958D2ee523a2206206994597C13D831ec7',
+            chains: {
+                'elated-tan-skat': {},
+                'honorable-steel-rasalhague': {
+                    hub: 'elated-tan-skat'
+                },
+                'green-giddy-denebola': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        usdp: {
+            address: '0x8E870D67F660D95d5be530380D0eC0bd388289E1',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        ruby: {
+            address: '0x918D8F3670c67f14Ff3fEB025D46B9C165d12a23',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        sushi: {
+            address: '0x6b3595068778dd592e39a122f4f5a5cf09c90fe2',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        link: {
+            address: '0x514910771af9ca656af840dff83e8264ecf986ca',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        pepe: {
+            address: '0x6982508145454ce325ddbe47a25d4ec3d2311933',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        cgt: {
+            address: '0x0E186357c323c806C1efdad36D217F7a54b63D18',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        wct1: {
+            address: '0x46683747B55C4A0fF783B1A502cE682eB819eb75',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        dai: {
+            address: '0x6B175474E89094C44Da98b954EedeAC495271d0F',
+            chains: {
+                'elated-tan-skat': {},
+                'honorable-steel-rasalhague': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        razor: {
+            address: '0x50DE6856358Cc35f3A9a57eAAA34BD4cB707d2cd',
+            chains: {
+                'elated-tan-skat': {},
+                'turbulent-unique-scheat': {}
+            }
+        },
+        usdc: {
+            address: '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48',
+            chains: {
+                'elated-tan-skat': {},
+                'honorable-steel-rasalhague': {
+                    hub: 'elated-tan-skat'
+                },
+                'green-giddy-denebola': {
+                    hub: 'elated-tan-skat'
+                },
+                'parallel-stormy-spica': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        hmt: {
+            address: '0xd1ba9BAC957322D6e8c07a160a3A8dA11A0d2867',
+            chains: {
+                'elated-tan-skat': {},
+                'wan-red-ain': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        exd: {
+            address: '0x02de007d412266a2e0fa9287c103474170f06560',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        hmkr: {
+            address: '0x3300b02efa180c99a2f61f4731665b51e4e254c4',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        paxg: {
+            address: '0x45804880De22913dAFE09f4980848ECE6EcbAf78',
+            chains: {
+                'elated-tan-skat': {}
+            }
+        },
+        skivvy: {
+            address: '0x246908BfF0b1ba6ECaDCF57fb94F6AE2FcD43a77',
+            chains: {
+                'elated-tan-skat': {},
+                'honorable-steel-rasalhague': {
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+    }
+}
+
+export default connections
