+++ config/mainnet/nebula.ts
@@ -1,0 +1,114 @@
+import { type types } from '@/core'
+
+const connections: types.mp.TokenTypeMap = {
+    eth: {
+        eth: {
+            address: '0xaB01BAd2C86e24D371A13eD6367bdCa819589C5D',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                },
+                'honorable-steel-rasalhague': {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        }
+    },
+    erc20: {
+        flag: {
+            address: '0x082081c8E607ca6C1c53aC093cAb3847ED59C0b0',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                }
+            }
+        },
+        reth: {
+            address: '0xa17dc9A0b3057F0c8169CEA7536B0f6bc46bc414',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        skl: {
+            address: '0x7F73B66d4e6e67bCdeaF277b9962addcDabBFC4d',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        usdc: {
+            address: '0xCC205196288B7A26f6D43bBD68AaA98dde97276d',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                },
+                'honorable-steel-rasalhague': {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                },
+                'parallel-stormy-spica': {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        usdt: {
+            address: '0x932427E1f1Ea096e2bb05C7dE937d083ddb8Ff83',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                },
+                'honorable-steel-rasalhague': {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        wbtc: {
+            address: '0x64d2EebA8B02f2aB19e2238f1655A8409dB64817',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                },
+                mainnet: {
+                    clone: true,
+                    hub: 'elated-tan-skat'
+                }
+            }
+        },
+        cmps: {
+            address: '0x3e1463814c83cD9dC25F1595314e62557588Ddde',
+            chains: {
+                'elated-tan-skat': {
+                    clone: true
+                }
+            }
+        }
+    }
+}
+
+export default connections
