+++ scripts/build.sh
@@ -7,11 +7,11 @@ set -e
 export DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 export PORTAL_DIR="$DIR/../"
 
-METAPORT_CONFIG_PATH=$PORTAL_DIR/config/$NETWORK_NAME.ts
-METAPORT_CONFIG_PATH_SRC=$PORTAL_DIR/src/data/metaportConfig.ts
+METAPORT_CONFIG_PATH=$PORTAL_DIR/config/$NETWORK_NAME
+METAPORT_CONFIG_PATH_SRC=$PORTAL_DIR/src/data/metaportConfig
 
 echo "Copying ${METAPORT_CONFIG_PATH} -> ${METAPORT_CONFIG_PATH_SRC}..."
-cp $METAPORT_CONFIG_PATH $METAPORT_CONFIG_PATH_SRC
+cp -R $METAPORT_CONFIG_PATH $METAPORT_CONFIG_PATH_SRC
 
 META_DIR_EXTERNAL=$PORTAL_DIR/skale-network/metadata/$NETWORK_NAME/
 META_DIR=$PORTAL_DIR/src/meta/
